+++ src/main/java/com/cinemaroyale/service/GeneroServiceImpl.java
@@ -4,6 +4,8 @@ import org.springframework.stereotype.Service;
 import com.cinemaroyale.dto.GeneroDTO;
 import com.cinemaroyale.model.Genero;
 import com.cinemaroyale.repository.GeneroRepository;
+import com.cinemaroyale.repository.PeliculaGeneroRepository;
+
 
 
 import java.util.List;
@@ -11,11 +13,13 @@ import java.util.stream.Collectors;
 
 @Service
 public class GeneroServiceImpl implements GeneroService {
+   private final GeneroRepository repository;
+    private final PeliculaGeneroRepository peliculaGeneroRepository;
 
-    private final GeneroRepository repository;
-
-    public GeneroServiceImpl(GeneroRepository repository) {
+    public GeneroServiceImpl(GeneroRepository repository,
+                             PeliculaGeneroRepository peliculaGeneroRepository) {
         this.repository = repository;
+        this.peliculaGeneroRepository = peliculaGeneroRepository;
     }
 
     private GeneroDTO mapToDTO(Genero g) {
@@ -59,6 +63,15 @@ public class GeneroServiceImpl implements GeneroService {
 
     @Override
     public void delete(Integer id) {
+
+        boolean enUso = peliculaGeneroRepository.existsByGeneroId(id);
+
+        if (enUso) {
+            throw new RuntimeException(
+                "No se puede eliminar el género porque está asignado a una película"
+            );
+        }
+
         repository.deleteById(id);
     }
 }
